requirements.txt
torch==2.6.0+cu124
torchvision==0.21.0+cu124
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.6.1